+++ stroom-receive/stroom-receive-common/src/main/java/stroom/receive/common/FeedStatusAttributeMapFilter.java
@@ -27,7 +27,7 @@ public class FeedStatusAttributeMapFilter implements AttributeMapFilter {
     public boolean filter(final AttributeMap attributeMap) {
         final String feedName = attributeMap.get(StandardHeaderArguments.FEED);
         if (feedName == null || feedName.trim().isEmpty()) {
-            throw new StroomStreamException(StroomStatusCode.FEED_MUST_BE_SPECIFIED);
+            throw new StroomStreamException(StroomStatusCode.FEED_MUST_BE_SPECIFIED, attributeMap);
         }
 
         final String senderDn = attributeMap.get(StandardHeaderArguments.REMOTE_DN);
@@ -35,7 +35,7 @@ public class FeedStatusAttributeMapFilter implements AttributeMapFilter {
         final GetFeedStatusResponse response = getFeedStatus(request);
 
         if (FeedStatus.Reject.equals(response.getStatus())) {
-            throw new StroomStreamException(StroomStatusCode.FEED_IS_NOT_SET_TO_RECEIVED_DATA);
+            throw new StroomStreamException(StroomStatusCode.FEED_IS_NOT_SET_TO_RECEIVED_DATA, attributeMap);
         }
 
         return FeedStatus.Receive.equals(response.getStatus());