+++ stroom-receive/stroom-receive-common/src/main/java/stroom/receive/common/ReceiveDataServlet.java
@@ -86,7 +86,7 @@ public class ReceiveDataServlet extends HttpServlet implements IsServlet {
             final RequestHandler requestHandler = requestHandlerProvider.get();
             requestHandler.handle(request, response);
         } catch (final RuntimeException e) {
-            StroomStreamException.sendErrorResponse(response, e);
+            StroomStreamException.sendErrorResponse(request, response, e);
         }
     }