+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
-Add debug logging to AuthenticationStateSessionUtil
+* Add debug logging to AuthenticationStateSessionUtil
+
+* Issue **#2448** : Change data receipt ERROR log messages to WARN. Also improve log message content for success and failure.
 
 * Issue **#2412** : You are now able to view locked streams or deleted data if it is still accessible.